+++ apps/shell/src/components/UserManager/index.ts
@@ -1,0 +1,3 @@
+export * from './UserManager.component';
+export * from './UserManager.hook';
+export * from './UserManager.types';